refactor: Split utils.py into multi-file module
- Analyzed 1283-line utils.py and identified logical components
- Created modular structure with 10 specialized files:
  - data_utils.py: DataFrame helpers and data snapshots
  - formatting.py: Price and volume formatters
  - market_utils.py: Market hours, session info, contract validation
  - trading_calculations.py: Tick values, position sizing, risk/reward
  - portfolio_analytics.py: Sharpe ratio, drawdown, portfolio metrics
  - pattern_detection.py: Candlestick and chart patterns
  - market_microstructure.py: Bid-ask spread and volume profile
  - rate_limiter.py: API rate limiting class
  - environment.py: Environment variable handling
  - logging_utils.py: Logging configuration
- Fixed all linting errors and sorted __all__ exports
- Tested all imports work correctly
- Maintained backward compatibility with existing imports
