+++ src/project_x_py/utils/environment.py
@@ -1,0 +1,25 @@
+"""Environment variable utilities."""
+
+import os
+from typing import Any
+
+
+def get_env_var(name: str, default: Any = None, required: bool = False) -> str:
+    """
+    Get environment variable with optional default and validation.
+
+    Args:
+        name: Environment variable name
+        default: Default value if not found
+        required: Whether the variable is required
+
+    Returns:
+        Environment variable value
+
+    Raises:
+        ValueError: If required variable is missing
+    """
+    value = os.getenv(name, default)
+    if required and value is None:
+        raise ValueError(f"Required environment variable '{name}' not found")
+    return value
