+++ src/project_x_py/utils/formatting.py
@@ -1,0 +1,16 @@
+"""Formatting utilities for prices, volumes, and other display values."""
+
+
+def format_price(price: float, decimals: int = 2) -> str:
+    """Format price for display."""
+    return f"${price:,.{decimals}f}"
+
+
+def format_volume(volume: int) -> str:
+    """Format volume for display."""
+    if volume >= 1_000_000:
+        return f"{volume / 1_000_000:.1f}M"
+    elif volume >= 1_000:
+        return f"{volume / 1_000:.1f}K"
+    else:
+        return str(volume)
