+++ src/project_x_py/utils/pattern_detection.py
@@ -1,0 +1,160 @@
+"""Pattern detection utilities for candlestick and chart patterns."""
+
+from typing import Any
+
+import polars as pl
+
+
+def detect_candlestick_patterns(
+    data: pl.DataFrame,
+    open_col: str = "open",
+    high_col: str = "high",
+    low_col: str = "low",
+    close_col: str = "close",
+) -> pl.DataFrame:
+    """
+    Detect basic candlestick patterns.
+
+    Args:
+        data: DataFrame with OHLCV data
+        open_col: Open price column
+        high_col: High price column
+        low_col: Low price column
+        close_col: Close price column
+
+    Returns:
+        DataFrame with pattern detection columns added
+
+    Example:
+        >>> patterns = detect_candlestick_patterns(ohlcv_data)
+        >>> doji_count = patterns.filter(pl.col("doji") == True).height
+        >>> print(f"Doji patterns found: {doji_count}")
+    """
+    required_cols = [open_col, high_col, low_col, close_col]
+    for col in required_cols:
+        if col not in data.columns:
+            raise ValueError(f"Column '{col}' not found in data")
+
+    # Calculate basic metrics
+    result = data.with_columns(
+        [
+            (pl.col(close_col) - pl.col(open_col)).alias("body"),
+            (pl.col(high_col) - pl.col(low_col)).alias("range"),
+            (pl.col(high_col) - pl.max_horizontal([open_col, close_col])).alias(
+                "upper_shadow"
+            ),
+            (pl.min_horizontal([open_col, close_col]) - pl.col(low_col)).alias(
+                "lower_shadow"
+            ),
+        ]
+    )
+
+    # Pattern detection
+    result = result.with_columns(
+        [
+            # Doji: Very small body relative to range
+            (pl.col("body").abs() <= 0.1 * pl.col("range")).alias("doji"),
+            # Hammer: Small body, long lower shadow, little upper shadow
+            (
+                (pl.col("body").abs() <= 0.3 * pl.col("range"))
+                & (pl.col("lower_shadow") >= 2 * pl.col("body").abs())
+                & (pl.col("upper_shadow") <= 0.1 * pl.col("range"))
+            ).alias("hammer"),
+            # Shooting Star: Small body, long upper shadow, little lower shadow
+            (
+                (pl.col("body").abs() <= 0.3 * pl.col("range"))
+                & (pl.col("upper_shadow") >= 2 * pl.col("body").abs())
+                & (pl.col("lower_shadow") <= 0.1 * pl.col("range"))
+            ).alias("shooting_star"),
+            # Bullish/Bearish flags
+            (pl.col("body") > 0).alias("bullish_candle"),
+            (pl.col("body") < 0).alias("bearish_candle"),
+            # Long body candles (strong moves)
+            (pl.col("body").abs() >= 0.7 * pl.col("range")).alias("long_body"),
+        ]
+    )
+
+    # Remove intermediate calculation columns
+    return result.drop(["body", "range", "upper_shadow", "lower_shadow"])
+
+
+def detect_chart_patterns(
+    data: pl.DataFrame,
+    price_column: str = "close",
+    window: int = 20,
+) -> dict[str, Any]:
+    """
+    Detect basic chart patterns.
+
+    Args:
+        data: DataFrame with price data
+        price_column: Price column to analyze
+        window: Window size for pattern detection
+
+    Returns:
+        Dict with detected patterns and their locations
+
+    Example:
+        >>> patterns = detect_chart_patterns(ohlcv_data)
+        >>> print(f"Double tops found: {len(patterns['double_tops'])}")
+    """
+    if price_column not in data.columns:
+        raise ValueError(f"Column '{price_column}' not found in data")
+
+    if len(data) < window * 2:
+        return {"error": "Insufficient data for pattern detection"}
+
+    try:
+        prices = data.select(pl.col(price_column)).to_series().to_list()
+
+        patterns = {
+            "double_tops": [],
+            "double_bottoms": [],
+            "breakouts": [],
+            "trend_reversals": [],
+        }
+
+        # Simple pattern detection logic
+        for i in range(window, len(prices) - window):
+            local_max = max(prices[i - window : i + window + 1])
+            local_min = min(prices[i - window : i + window + 1])
+            current_price = prices[i]
+
+            # Double top detection (simplified)
+            if current_price == local_max:
+                # Look for another high nearby
+                for j in range(i + window // 2, min(i + window, len(prices))):
+                    if (
+                        abs(prices[j] - current_price) / current_price < 0.02
+                    ):  # Within 2%
+                        patterns["double_tops"].append(
+                            {
+                                "index1": i,
+                                "index2": j,
+                                "price": current_price,
+                                "strength": local_max - local_min,
+                            }
+                        )
+                        break
+
+            # Double bottom detection (simplified)
+            if current_price == local_min:
+                # Look for another low nearby
+                for j in range(i + window // 2, min(i + window, len(prices))):
+                    if (
+                        abs(prices[j] - current_price) / current_price < 0.02
+                    ):  # Within 2%
+                        patterns["double_bottoms"].append(
+                            {
+                                "index1": i,
+                                "index2": j,
+                                "price": current_price,
+                                "strength": local_max - local_min,
+                            }
+                        )
+                        break
+
+        return patterns
+
+    except Exception as e:
+        return {"error": str(e)}
